# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Outpost VPN API
  description: |
    Open-source WireGuard VPN & Zero-Trust Network Access platform.
    All features available under Apache 2.0 — no enterprise paywalls.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Romashqua Labs
    url: https://github.com/romashqua/outpost
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://vpn.example.com
    description: Production
...
paths:
...
  /api/v1/scim/v2/Users:
    get:
      operationId: scimListUsers
      summary: List users (SCIM format)
      tags: [SCIM]
      parameters:
        - name: startIndex
          in: query
          schema:
            type: integer
            default: 1
        - name: count
          in: query
          schema:
            type: integer
            default: 100
        - name: filter
          in: query
          description: "SCIM filter (e.g. userName eq \"john\")"
          schema:
            type: string
      responses:
        "200":
          description: SCIM user list
          content:
            application/scim+json:
              schema:
                $ref: "#/components/schemas/SCIMListResponse"

    post:
      operationId: scimCreateUser
      summary: Create user via SCIM
      tags: [SCIM]
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: "#/components/schemas/SCIMUser"
      responses:
        "201":
          description: User created
          content:
            application/scim+json:
              schema:
                $ref: "#/components/schemas/SCIMUser"
        "409":
          description: User already exists
          content:
            application/scim+json:
              schema:
                $ref: "#/components/schemas/SCIMError"
  /api/v1/scim/v2/Users/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceID"

    get:
      operationId: scimGetUser
      summary: Get user via SCIM
      tags: [SCIM]
      responses:
        "200":
          description: SCIM user
          content:
            application/scim+json:
              schema:
                $ref: "#/components/schemas/SCIMUser"
        "404":
          description: User not found
          content:
            application/scim+json:
              schema:
                $ref: "#/components/schemas/SCIMError"

    put:
      operationId: scimReplaceUser
      summary: Replace user via SCIM
      tags: [SCIM]
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              $ref: "#/components/schemas/SCIMUser"
      responses:
        "200":
          description: User replaced
          content:
            application/scim+json:
              schema:
                $ref: "#/components/schemas/SCIMUser"

    patch:
      operationId: scimUpdateUser
      summary: Patch user via SCIM
      tags: [SCIM]
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              type: object
              properties:
                schemas:
                  type: array
                  items:
                    type: string
                Operations:
                  type: array
                  items:
                    type: object
                    properties:
                      op:
                        type: string
                        enum: [add, replace, remove]
                      path:
                        type: string
                      value: {}
      responses:
        "200":
          description: User updated
          content:
            application/scim+json:
              schema:
                $ref: "#/components/schemas/SCIMUser"

    delete:
      operationId: scimDeleteUser
      summary: Deactivate user via SCIM
      tags: [SCIM]
      responses:
        "204":
          description: User deactivated
...
  /api/v1/users:
    get:
      operationId: listUsers
      summary: List all users
      tags: [Users]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"

    post:
      operationId: createUser
      summary: Create a new user
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/users/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceID"

    get:
      operationId: getUser
      summary: Get user by ID
      tags: [Users]
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      operationId: updateUser
      summary: Update a user
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      operationId: deleteUser
      summary: Deactivate a user (soft delete)
      tags: [Users]
      responses:
        "200":
          description: User deactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: deactivated
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Networks ────────────────────────────────────────────────────────
...
  /api/v1/users/{id}/activate:
    patch:
      operationId: activateUser
      summary: Reactivate a deactivated user
      tags: [Users]
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: User reactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid UUID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Device Enrollment ────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained via POST /api/v1/auth/login
  parameters:
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        default: 1
        minimum: 1
    PerPage:
      name: per_page
      in: query
      description: Items per page
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 100
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: something went wrong

    # ── Auth ────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, username, email, is_active, is_admin, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
          nullable: true
        is_active:
          type: boolean
        is_admin:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserListResponse:
      type: object
      required: [data, page, per_page]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        page:
          type: integer
        per_page:
          type: integer
    CreateUserRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        first_name:
          type: string
        last_name:
          type: string
        is_admin:
          type: boolean
          default: false
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        is_active:
          type: boolean
        is_admin:
          type: boolean

    # ── Networks ────────────────────────────────────────────────────
...
    SCIMListResponse:
      type: object
      required: [schemas, totalResults, Resources]
      properties:
        schemas:
          type: array
          items:
            type: string
          example: ["urn:ietf:params:scim:api:messages:2.0:ListResponse"]
        totalResults:
          type: integer
        startIndex:
          type: integer
        itemsPerPage:
          type: integer
        Resources:
          type: array
          items:
            type: object
    SCIMUser:
      type: object
      required: [schemas, userName]
      properties:
        schemas:
          type: array
          items:
            type: string
          example: ["urn:ietf:params:scim:schemas:core:2.0:User"]
        id:
          type: string
          format: uuid
        externalId:
          type: string
        userName:
          type: string
        name:
          type: object
          properties:
            givenName:
              type: string
            familyName:
              type: string
        emails:
          type: array
          items:
            type: object
            properties:
              value:
                type: string
                format: email
              primary:
                type: boolean
        active:
          type: boolean
        meta:
          type: object
          properties:
            resourceType:
              type: string
              example: User
            created:
              type: string
              format: date-time
            lastModified:
              type: string
              format: date-time
...
    SCIMError:
      type: object
      required: [schemas, status, detail]
      properties:
        schemas:
          type: array
          items:
            type: string
          example: ["urn:ietf:params:scim:api:messages:2.0:Error"]
        status:
          type: string
          example: "409"
        detail:
          type: string

    # ── Dashboard ────────────────────────────────────────────────────
...
security:
  - BearerAuth: []